refactor: api 디렉토리 위치 src/components/api -> src/api로 이동
